deal with multi input tensors (#99)
* deal with multi input tensors

* fix input order successful

* format and refine

* refine

* support global tensor input

* refine

* add input_tensor_range and log info

* refine

* fix comment

* add onnx_simplify

* code format

* fix add_n bug

* refine

* refine

+++ oneflow_onnx/oneflow2onnx/handlers/array.py
@@ -340,7 +340,7 @@ class Constant:
         if is_floating_value:
             values = np.full(shape=shape, fill_value=floating_value, dtype=np.float32)
         else:
-            values = np.full(shape=shape, fill_value=integer_value, dtype=np.float32)
+            values = np.full(shape=shape, fill_value=integer_value, dtype=np.int64)
         output_name = node.output_tensor_names[0]
         ctx.RemoveNode(node.name)
         if is_floating_value:
